Provide a lightweight client for use in App Engine
One cannot keep a Client around in App Engine. Instead, you must create
a new Client for every request. We cannot afford to keep Goroutines
around and must make creation of a Client as efficient as possible.

This fix removes many of the sanity checks in `NewClient` when the
sniffer and the healthchecker are disabled via `SetSniff(false)` and
`SetHealthcheck(false)`.